+++ sdk/purview/azure-analytics-purview-administration/src/samples/java/com/azure/analytics/purview/administration/GetAccount.java
@@ -8,10 +8,10 @@ import com.azure.identity.DefaultAzureCredentialBuilder;
 
 public class GetAccount {
     public static void main(String[] args) {
-        AccountsClient client = new PurviewAccountClientBuilder()
+        AccountsClient client = new AccountsClientBuilder()
             .endpoint(System.getenv("ACCOUNT_ENDPOINT"))
             .credential(new DefaultAzureCredentialBuilder().build())
-            .buildAccountsClient();
+            .buildClient();
         BinaryData response = client.getAccountPropertiesWithResponse(null).getValue();
         System.out.println(response);
     }